requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
pandas
plotly
langchain-community
# -e .